+++ spec/dumpable_spec.rb
@@ -24,7 +24,7 @@ describe 'Dumpable' do
      }
 
     it 'do not get the base exception' do
-      expect{ type.new.dump }.to_not raise_exception(NotImplementedError)
+      expect{ type.new.dump }.to_not raise_exception
     end
   end
 end